# Reorganize documentation: reduce clutter by 55%, improve discoverability